Giant reworking of navigation map
Now returns either all the best options, or just the first (hopefully random, no longer favors straight directions, which should be done upstream).  Includes a lambda to filter the options, which allows consumers to better tune which directions are considered.  For example, the map might be initiated to ignore creatures, but skip them when computing the best next move.